reorganize tasks in separate celery apps
